backend/requirements.txt
# ============================================
# NexInvo Backend Dependencies
# Last Updated: December 2024
# ============================================

# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
django-extensions==4.1

# API Documentation
drf-spectacular==0.28.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
cryptography==42.0.2
pyotp==2.9.0

# Database
psycopg2-binary==2.9.11
psycopg==3.2.12
psycopg-binary==3.2.12
django-redis==6.0.0
redis==5.0.1

# PDF Generation & Processing
reportlab==4.4.4
Pillow==11.3.0
PyMuPDF==1.26.6
pypdf==6.1.3
PyPDF2==3.0.1
pdfminer.six==20251107
# docx2pdf==0.1.8  # Windows-only (requires Microsoft Word COM)

# Excel/CSV/Document Processing
openpyxl==3.1.2
pandas==2.3.3
python-docx==1.2.0
htmldocx==0.0.6
lxml==6.0.2
pypandoc==1.16

# QR Code Generation
qrcode==8.2

# Background Job Scheduling
django-apscheduler==0.7.0
APScheduler==3.11.1
celery==5.3.6
kombu==5.5.4
amqp==5.3.1
billiard==4.2.2
vine==5.1.0

# Google Services (Gmail API, etc.)
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
googleapis-common-protos==1.72.0
google-api-core==2.25.2

# HTTP & Networking
requests==2.32.4
httplib2==0.31.0
urllib3==2.5.0
websocket-client==1.9.0
websockets==15.0.1

# Web Scraping & Automation
selenium==4.38.0
webdriver-manager==4.0.2
undetected-chromedriver==3.5.5
beautifulsoup4==4.14.2
soupsieve==2.8

# Data Validation & Parsing
pydantic==2.12.4
pydantic_core==2.41.5

# Utilities
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
tzdata==2025.2
tzlocal==5.3.1
six==1.17.0

# CLI & Formatting
click==8.3.0
click-default-group==1.2.4
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
tabulate==0.9.0
tqdm==4.67.1
colorama==0.4.6

# Monitoring & Error Tracking
sentry-sdk==2.44.0
psutil==7.1.3

# Production Server
gunicorn==23.0.0
whitenoise==6.11.0

# Protocol Buffers & gRPC
protobuf==5.29.5
proto-plus==1.26.1
grpcio==1.76.0
grpcio-status==1.71.2

# Calendar & Scheduling
icalendar==6.3.2

# Flask (for specific tools/utilities)
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Werkzeug==3.1.4
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0

# WebSocket Support
python-socketio==5.15.0
python-engineio==4.12.3
simple-websocket==1.1.0

# Django Channels (for Setu WebSocket integration)
channels==4.2.0
channels-redis==4.2.1
daphne==4.2.0

# Testing & Load Testing
pytest==9.0.2
locust==2.42.6
locust-cloud==1.29.5

# Async/Concurrent Processing
gevent==25.9.1
geventhttpclient==2.3.7
greenlet==3.3.0

# Additional Utilities
babel==2.17.0
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.1
ConfigArgParse==1.7.1
h11==0.16.0
iniconfig==2.3.0
keyring==25.6.0
msgpack==1.1.2
oauthlib==3.3.1
outcome==1.3.0.post0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
Pygments==2.19.2
PySocks==1.7.1
pyzmq==27.1.0
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlite-fts4==1.0.3
sqlite-utils==3.38
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
uritemplate==4.2.0
wcwidth==0.2.14
webencodings==0.5.1
wsproto==1.2.0
zope.event==6.1
zope.interface==8.1.1

# Desktop GUI (Windows-only - DO NOT install on Linux servers)
# Uncomment these only for Windows desktop development
# pywin32==311
# pypiwin32==223
# pywin32-ctypes==0.2.3
# tkcalendar==1.6.1
# tkinterdnd2==0.4.3
# ttkbootstrap==1.18.1
# plyer==2.1.0
# win10toast==0.9

# Core Python Dependencies
asgiref==3.10.0
sqlparse==0.5.3
numpy==2.3.4
matplotlib==3.10.7
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.0
kiwisolver==1.4.9
pyparsing==3.2.5
bleach==6.1.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
packaging==25.0
typing_extensions==4.15.0
annotated-types==0.7.0
et_xmlfile==2.0.0
cffi==2.0.0
pycparser==2.23
pyasn1==0.6.1
pyasn1_modules==0.4.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
more-itertools==10.8.0
